requirements.txt
flask
flask-cors
flask-sqlalchemy
gunicorn
pandas
psycopg2-binary
python-decouple
requests
scikit-learn
